Move tests from GeoHashNtiveTest to GeoHashesTest
